+++ tests/test_validation.py
@@ -1,7 +1,9 @@
 import eido
 import pytest
 from pephub.dependencies import *
-pytest_plugins = ('pytest_asyncio',)
+
+pytest_plugins = ("pytest_asyncio",)
+
 
 # test project file on schema file validation success
 def test_file_file_validate_valid(project_object_file, schema_file_path):
@@ -11,7 +13,9 @@ def test_file_file_validate_valid(project_object_file, schema_file_path):
 # test project file on schema file validation failure
 def test_file_file_validate_invalid(project_object_file, schema_file_path_invalid):
     with pytest.raises(eido.exceptions.EidoValidationError):
-        eido.validate_project(project=project_object_file, schema=schema_file_path_invalid)
+        eido.validate_project(
+            project=project_object_file, schema=schema_file_path_invalid
+        )
 
 
 # test project file on schema string validation success
@@ -33,7 +37,9 @@ def test_file_url_validate_valid(project_object_file, schema_from_url_valid):
 # test project file on schema url validation failure
 def test_file_url_validate_invalid(project_object_file, schema_from_url_invalid):
     with pytest.raises(eido.exceptions.EidoValidationError):
-        eido.validate_project(project=project_object_file, schema=schema_from_url_invalid)
+        eido.validate_project(
+            project=project_object_file, schema=schema_from_url_invalid
+        )
 
 
 def test_registry_paste_valid(db, schema_paste):
@@ -67,6 +73,3 @@ def test_registry_url_invalid(db, schema_from_url_invalid):
     p = db.project.get("ayobi", "new-project-test12345", "default")
     with pytest.raises(eido.exceptions.EidoValidationError):
         eido.validate_project(project=p, schema=schema_from_url_invalid)
-
-
-