+++ docs/changelog.md
@@ -2,16 +2,37 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.9.4] - 07-18-2023
+
+### Fixed
+
+- Typo in tooltip for search bar
+
+### Added
+
+- Tooltip on landing page namespace list
+
+### Changed
+
+- Styling of landing namespaces to more clearly indicate they are links
+
+## [0.9.3] - 07-17-2023
+
+### Changed
+
+- Styling updates
+- Landing tooltips
+- Minor UI updates
+
 ## [0.9.2] - 07-12-2023
 
 ### Fixed
 
 - validating was not firing when updating sample table, subsample table, or config
 
-
 ### Added
 
-- github organizations UI visibility 
+- github organizations UI visibility
 - schema tag has link to schema
 
 ## [0.9.1] - 07-11-2023
@@ -21,11 +42,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - forking was broken
 - order in config file was incorrect
 
-
 ### Added
 
 - config endpoint
-
 
 ## [0.9.0] - 07-05-2023